finished editUserMockup
