+++ .gitignore
@@ -9,9 +9,9 @@
 
 # Don't commit the production settings file.
 settings.production.json
+settings.json
 
 # Build node_modules locally, ignore errors.
 node_modules/
 npm-debug.log
 
-config/settings.json
