+++ app/client/main.js
@@ -16,3 +16,4 @@ import '/imports/ui/stylesheets/style.css';
 import '/imports/api/base';
 import '/imports/api/profile';
 import '/imports/api/interest';
+
