+++ app/imports/ui/components/user/interests-form-field.html
@@ -3,11 +3,11 @@
     <label>Interest(s)</label>
     <select multiple="" class="ui dropdown" name="interests">
       <option value="">Select Interest(s)</option>
-        <option value="algorithms">Algorithms</option>
-        <option value="software-engineering">Software Engineering</option>
-        <option value="scuba">Scuba</option>
-        <option value="accreditation">Accreditation</option>
-        <option value="surfing">Surfing</option>
+        <option value="algorithms">Pasta</option>
+        <option value="software-engineering">Pizza</option>
+        <option value="scuba">Salad</option>
+        <option value="accreditation">Seafood</option>
+        <option value="surfing">Chicken</option>
     </select>
   </div>
 </template>
