+++ app/imports/ui/layouts/user/user-header.html
@@ -4,6 +4,7 @@
       <div class="item"><h1>ManoaStudentKitchen</h1></div>
       <a class="{{isActiveRoute "Home_Page"}} item" href="{{pathFor "Home_Page" username=routeUserName}}">Home</a>
       <a class="{{isActiveRoute "Profile_Page"}} item" href="{{pathFor "Profile_Page" username=routeUserName}}">Profile</a>
+      <a class="{{isActiveRoute "Edit_Profile_Page"}} item" href="{{pathFor "Edit_Profile_Page" username=routeUserName}}">Edit Profile</a>
       <a class="{{isActiveRoute "Filter_Page"}} item" href="{{pathFor "Filter_Page" username=routeUserName}}">Filter</a>
       <div class="right aligned item">{{> Cas_Login }}</div>
     </div>
