+++ app/imports/ui/pages/home/home-page.html
@@ -20,7 +20,7 @@
     <div class="ui link centered cards">
       <div class="card">
         <div class="ui centered card-image">
-          <img class="ui centered image" src="images/recipe.png">
+          <img class="ui centered image" src="/images/recipe.png">
         </div>
         <div class="content">
           <div class="header">Recipe</div>
@@ -35,20 +35,20 @@
         </div>
 
         <div class="extra content">
-
-
-        <span class="right floated">
-
-        </span>
           <span>
+            <div class="ui label">
+              Likes
+              <div class="detail">[###]</div>
+            </div>
+          </span>
+          <span class="right floated">
             Added in 2017
-           </span>
-
+          </span>
         </div>
       </div>
       <div class="card">
         <div class="ui centered card-image">
-          <img class="ui centered image" src="images/recipe.png">
+          <img class="ui centered image" src="/images/recipe.png">
         </div>
         <div class="content">
           <div class="header">Recipe</div>
@@ -63,20 +63,20 @@
         </div>
 
         <div class="extra content">
-
-
-        <span class="right floated">
-
-        </span>
           <span>
+            <div class="ui label">
+              Likes
+              <div class="detail">[###]</div>
+            </div>
+          </span>
+          <span class="right floated">
             Added in 2017
-           </span>
-
+          </span>
         </div>
       </div>
       <div class="card">
         <div class="ui centered card-image">
-          <img class="ui centered image" src="images/recipe.png">
+          <img class="ui centered image" src="/images/recipe.png">
         </div>
         <div class="content">
           <div class="header">Recipe</div>
@@ -91,15 +91,15 @@
         </div>
 
         <div class="extra content">
-
-
-        <span class="right floated">
-
-        </span>
           <span>
+            <div class="ui label">
+              Likes
+              <div class="detail">[###]</div>
+            </div>
+          </span>
+          <span class="right floated">
             Added in 2017
-           </span>
-
+          </span>
         </div>
       </div>
