+++ app/imports/ui/pages/user/index.js
@@ -1,2 +1,4 @@
 import './profile-page.html';
-import './profile-page';
+import './profile-page.js';
+import './edit-profile-page.html';
+import './edit-profile-page.js';
