+++ app/private/database/initial-collection-data.json
@@ -5,8 +5,8 @@
       "name": "InterestCollection",
       "contents": [
         {
-          "name": "Software Engineering",
-          "description": "Techniques for development of high quality systems"
+          "name": "Pizza",
+          "description": ""
         },
         {
           "name": "Algorithms",
